Refactor About section into Settings screen
Fixes #639

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/enaboapps/Switchify/issues/639?shareId=XXXX-XXXX-XXXX-XXXX).
